The 3D models were organized by components
