+++ src/pages/world/World.jsx
@@ -1,6 +1,7 @@
 import { OrbitControls } from "@react-three/drei";
 import { Canvas } from "@react-three/fiber";
-import House from "./House";
+import House3d from "../../components/House3d/House3d";
+import TestDummy3d from "../../components/TestDummy/TestDummy3d";
 
 const World = () => { 
     return <Canvas
@@ -13,7 +14,8 @@ const World = () => {
         <ambientLight intensity={1.5}/>
         <directionalLight position={[0,10,10]}/>
         <OrbitControls/>
-        <House/>
+        <House3d />
+        <TestDummy3d />
     </Canvas>;
     
 };
